requirements.txt
python-3.9.18
numpy-1.26.2
pytorch-2.1.2
pytorch-cuda-12.1
tenseal-0.3.14
gmpy2-2.1.5
gap-stat-2.0.3
pillow-6.1
tenseal-0.3.14
scikit-learn-1.3.0
pandas-2.1.4